Added a guided missile
I added a prefab with script and a particle effect.  The missile will track the player for X seconds and then accelerate in its current direction after X seconds (no player tracking).  It register a hit and removes it self but does not damage the player right now.  Needs an explosion.